refactor(codegen): split static-cache emission into children_static
Move the static-subtree cache emitters (`is_static_cacheable_element`,
`generate_cached_static_children_array`, `generate_cached_static_element`,
`generate_cached_static_vnode`) out of `children.rs` into a new
`children_static` module, so the children dispatch file stays focused on
grouping and the Rendu op stream. Pure code movement — no behavior change.

This drops `children.rs` from 351 to 265 lines, back under the source-length
guard, making room to route `generate_children_inner` through the Rendu op
stream next (#1756). Verified byte-equivalent across vize_atelier_dom,
vize_atelier_sfc (528), and vize_atelier_ssr (76); clippy and rustfmt clean.
